% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{my_sample}
\alias{my_sample}
\title{description about function.}
\usage{
my_sample(x)
}
\arguments{
\item{x}{descriton about input value}
}
\value{
descrtion about result value
}
\description{
description about function.
}
